match function names to libdragon preview

+++ src/pak_rom.h
@@ -1,6 +1,8 @@
 #ifndef TORTIMER_PAK_ROM_H
 #define TORTIMER_PAK_ROM_H
 
+#include <stddef.h>
+
 /**
  * Writes the ROM data to the Controller Pak.
  * @param data Pointer to the ROM data.
@@ -9,7 +11,7 @@
  * @param compress If non-zero, compress the ROM data using Yay0.
  * @return 0 on success, negative error code on failure.
  */
-int pak_write_rom(unsigned char *data, unsigned size, const char *name,
+int pak_write_rom(unsigned char *data, size_t size, const char *name,
   int compress);
 
 #endif
